requirements.txt
asgiref==3.5.2
axios==0.1.0
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==38.0.4
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.51.0
django-bootstrap5==22.2
django-cleanup==6.0.0
django-extensions==3.2.1
django-summernote==0.8.20.0
django-widget-tweaks==1.4.12
idna==3.4
lxml==4.9.1
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
webencodings==0.5.1
